Add app to reintegrate with VS Code
When the project was originally created, it had a console app that was used by VS Code to run/debug things. That became the class library under src, so let's add ti-cli back as the app to actually run things.

We'll move the Unit class from where the interfaces are defined into the Program class to actually execute some code when running.